Add dataset definition for diabetes register
This adds the dataset definition for diabetes register (DM_REG).
I'm expecting to add more dataset definitions for further indicators and am therefore adding a suffix (*_dm_reg) to all files and outputs.
Currently all variables that are needed are defined within this dataset definition for the register.
This is ok for now, but we might want to change this later because some of the variables are shared across different indicators and it would therefore make sense to define these variables in a variable dictionary that can be called in different dataset definitions.