Add indicator dm020 vars and dataset definition
All variables needed for dm020 as well as the business logic get defined in `variable_lib_dm.py`.
Because the variables for indicators dm020 and dm021 are identical logic but different cutoff values for `ifcchba_val` the same functions can be used for both indicators (the naming of the functions is not ideal but I decided to be descriptive).
In each dataset definition the variables first get created using the functions from the variable library.
Second, the business rule logic gets specified.
Each select rule in the business rules gets specified and combines with an OR statement in dataset.set_population().