Generate dataset and measures for dm017
We checks if two dates are provided in project.yaml and if
so assigns them to be the start and end date of the interval.
This allows us to generate a dataset and measures from the same dataset
definition. We also need to do the same check for .define_population()
because the measrues framework returns an error if .define_population() gets used.